finish basic wave spawning

+++ include/enemies.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <stdint.h>
 #include "player.h"
 #include "projectile.h"
 
@@ -9,4 +9,4 @@ extern uint32_t numEnemies;
 void damageEnemy(uint32_t enemyIndex, uint32_t damage);
 void updateEnemies();
 void drawEnemies();
-void spawnEnemy(float rotationSpeed, float attackSpeed, float maxHealth, float moveSpeed, Color color, Vector2 position);
+void spawnEnemy(float rotationSpeed, float attackSpeed, uint32_t maxHealth, float moveSpeed, Color color, Vector2 position);
